Migrated SetContentType tests.
